requirements.txt
flask==2.3.3
tensorflow==2.15.0
numpy==1.23.5
Pillow==10.1.0
werkzeug==3.0.1
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2